Room database

+++ app/src/main/java/at/tamburi/tamburimontageservice/models/ServiceUser.kt
@@ -3,7 +3,7 @@ package at.tamburi.tamburimontageservice.models
 data class ServiceUser(
     val userId: Int,
     val username: String,
-    val loginDate: Int,
+    val loginDate: Long,
     val magazineId: Int?,
     val assignedMontageTaskId: Int?
     )
